Make initial commit.
This Vim plugin remaps keys on a Colemak keyboard/layout to their QWERTY
equivalent so that commands use the same physical keystrokes.  Only
command keystrokes (e.g., CTRL-R) are mapped in Insert and Command-line
modes, so normal text is entered with Colemak.

CTRL-P is always remapped to <Leader>; because CTRL-; doesn't correspond
to a valid ASCII control character.  Commands in CTRL-X submode are not
mapped because I don't think they can be.  There should be some kind of
real documentation, but there isn't yet.  Et cetera.