Switch to a two-step build process
Until now I'd been using a single-step build process, treating the
destination dotfiles as Make targets and writing M4 output to them
directly. This was elegant and simple (to me at least), but it also
imposed limitations on the destination path. Notably, said path could
not contain spaces because Make targets are space-delimited. This was
never a problem in practice, but it felt sloppy.

A more-conventional two-step build process (`make && make install`)
partially addresses this issue. The first step writes the M4 output to
files within the repository checkout, where I can simply avoid spaces in
filenames. The second step installs those files to the user-selected
destination using phony targets; it avoids referencing the destination
path in places where spaces are a problem.

Although the build system itself is now compatible with spaces in
pathnames, the templates still require their own adjustments.

As a bonus, the two-step process inherently allows the installation of
files that are not templates. Any entry in the *_files lists that does
not have a corresponding *.m4 file is copied directly without being run
through M4. This is useful for binary files.

Minor notes:
- The build system no longer uses directory search, so stop calling the
  module list VPATH.
- The M4 templates' names will never contain spaces or symbols, so don't
  bother quoting them in the recipe.